requirements.txt
# RAG with Gemma and PDF Files

## Core Dependencies
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=1.0.0
chromadb>=0.5.0
pypdf>=5.0.0
python-dotenv>=1.0.0

## Optional Dependencies
# For better PDF parsing
pymupdf>=1.24.0

# For Arabic text processing
arabic-reshaper>=2.1.0
python-bidi